assignment-analyser-backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pymupdf
openai
python-dotenv
reportlab
